requirements.txt
streamlit
python-dotenv
llama-index
PyMuPDF
pytesseract
Pillow
chromadb
